chat_app/app/requirements.txt
fastapi
uvicorn
jinja2
pydantic
sqlalchemy
python-dotenv
openai
langsmith
requests
aiofiles
async-exit-stack
async-generator
starlette

